+++ modules/apps/headless/headless-delivery/headless-delivery-impl/src/main/java/com/liferay/headless/delivery/internal/dto/v1_0/mapper/ContainerLayoutStructureItemMapper.java
@@ -23,16 +23,16 @@ import com.liferay.headless.delivery.dto.v1_0.PageSectionDefinition;
 import com.liferay.headless.delivery.internal.dto.v1_0.mapper.util.FragmentMappedValueUtil;
 import com.liferay.headless.delivery.internal.dto.v1_0.mapper.util.LocalizedValueUtil;
 import com.liferay.headless.delivery.internal.dto.v1_0.mapper.util.StyledLayoutStructureItemUtil;
-import com.liferay.layout.util.converter.AlignConverter;
-import com.liferay.layout.util.converter.BorderRadiusConverter;
-import com.liferay.layout.util.converter.ContentDisplayConverter;
-import com.liferay.layout.util.converter.ContentVisibilityConverter;
-import com.liferay.layout.util.converter.FlexWrapConverter;
-import com.liferay.layout.util.converter.HtmlTagConverter;
-import com.liferay.layout.util.converter.JustifyConverter;
-import com.liferay.layout.util.converter.MarginConverter;
-import com.liferay.layout.util.converter.PaddingConverter;
-import com.liferay.layout.util.converter.ShadowConverter;
+import com.liferay.layout.converter.AlignConverter;
+import com.liferay.layout.converter.BorderRadiusConverter;
+import com.liferay.layout.converter.ContentDisplayConverter;
+import com.liferay.layout.converter.ContentVisibilityConverter;
+import com.liferay.layout.converter.FlexWrapConverter;
+import com.liferay.layout.converter.HtmlTagConverter;
+import com.liferay.layout.converter.JustifyConverter;
+import com.liferay.layout.converter.MarginConverter;
+import com.liferay.layout.converter.PaddingConverter;
+import com.liferay.layout.converter.ShadowConverter;
 import com.liferay.layout.util.structure.CommonStylesUtil;
 import com.liferay.layout.util.structure.ContainerStyledLayoutStructureItem;
 import com.liferay.layout.util.structure.LayoutStructureItem;
