+++ modules/apps/layout/layout-impl/src/main/java/com/liferay/layout/internal/importer/structure/util/ContainerLayoutStructureItemImporter.java
@@ -16,17 +16,17 @@ package com.liferay.layout.internal.importer.structure.util;
 
 import com.liferay.headless.delivery.dto.v1_0.FragmentLink;
 import com.liferay.headless.delivery.dto.v1_0.PageElement;
+import com.liferay.layout.converter.AlignConverter;
+import com.liferay.layout.converter.BorderRadiusConverter;
+import com.liferay.layout.converter.ContentDisplayConverter;
+import com.liferay.layout.converter.ContentVisibilityConverter;
+import com.liferay.layout.converter.FlexWrapConverter;
+import com.liferay.layout.converter.HtmlTagConverter;
+import com.liferay.layout.converter.JustifyConverter;
+import com.liferay.layout.converter.MarginConverter;
+import com.liferay.layout.converter.PaddingConverter;
+import com.liferay.layout.converter.ShadowConverter;
 import com.liferay.layout.internal.importer.LayoutStructureItemImporterContext;
-import com.liferay.layout.util.converter.AlignConverter;
-import com.liferay.layout.util.converter.BorderRadiusConverter;
-import com.liferay.layout.util.converter.ContentDisplayConverter;
-import com.liferay.layout.util.converter.ContentVisibilityConverter;
-import com.liferay.layout.util.converter.FlexWrapConverter;
-import com.liferay.layout.util.converter.HtmlTagConverter;
-import com.liferay.layout.util.converter.JustifyConverter;
-import com.liferay.layout.util.converter.MarginConverter;
-import com.liferay.layout.util.converter.PaddingConverter;
-import com.liferay.layout.util.converter.ShadowConverter;
 import com.liferay.layout.util.structure.ContainerStyledLayoutStructureItem;
 import com.liferay.layout.util.structure.LayoutStructure;
 import com.liferay.layout.util.structure.LayoutStructureItem;
